README.md
A basic Tetris game created as part of my journey to becoming a game developer. This project is one of many stepping stones towards developing my first full game, allowing me to learn and practice key game development concepts. I'm sharing this with the community to document my progress and to inspire others who are also on a similar path.